requirements.txt
asgiref==3.8.1
azure-core==1.32.0
azure-iot-hub==2.6.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
isodate==0.7.2
msrest==0.7.1
oauthlib==3.2.2
psycopg2==2.9.10
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uamqp==1.6.11
urllib3==2.2.3
whitenoise==6.8.2
